add skeleton + refacto
