+++ conversation/frontend/src/features/absence/hooks/index.ts
@@ -1,2 +1,2 @@
 export * from './useAbsenceModal';
-export * from './useAbsenceModalContainer';
+export * from './useAbsenceModalData';
