+++ conversation/frontend/src/features/absence/hooks/useAbsenceModalData.test.ts
@@ -3,15 +3,15 @@ import { mockAbsenceSettings } from '~/mocks';
 import { wrapper } from '~/mocks/setup';
 import { queryClient } from '~/providers';
 import { absenceService } from '~/services';
-import { useAbsenceModalContainer } from './useAbsenceModalContainer';
+import { useAbsenceModalData } from './useAbsenceModalData';
 
-describe('useAbsenceModalContainer', () => {
+describe('useAbsenceModalData', () => {
   beforeEach(() => {
     queryClient.clear();
   });
 
   it('starts loading, then settles with the fetched settings', async () => {
-    const { result } = renderHook(useAbsenceModalContainer, { wrapper });
+    const { result } = renderHook(useAbsenceModalData, { wrapper });
 
     expect(result.current.isLoadingSettings).toBe(true);
     expect(result.current.settings).toBeUndefined();
@@ -28,7 +28,7 @@ describe('useAbsenceModalContainer', () => {
 
   it('saves through the mutation and reflects its pending state', async () => {
     const saveSpy = vi.spyOn(absenceService, 'saveSettings');
-    const { result } = renderHook(useAbsenceModalContainer, { wrapper });
+    const { result } = renderHook(useAbsenceModalData, { wrapper });
     await waitFor(() => expect(result.current.isLoadingSettings).toBe(false));
 
     const payload = {
