+++ conversation/frontend/src/features/index.ts
@@ -1,5 +1,4 @@
 export * from './absence/AbsenceModal';
-export * from './absence/AbsenceModalContainer';
 export * from './menu';
 export * from './modals/CreateFolderModal';
 export * from './modals/RenameFolderModal';
